% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchess.R
\docType{package}
\name{rchess-package}
\alias{rchess}
\alias{rchess-package}
\title{Chess tools for R}
\description{
Chess move generation and validation, piece placement and movement,
game-state inspection, and board visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jkunst.com/rchess/}
  \item \url{https://github.com/jbkunst/rchess}
  \item Report bugs at \url{https://github.com/jbkunst/rchess/issues}
}

}
\author{
\strong{Maintainer}: Joshua Kunst \email{jbkunst@gmail.com}

Authors:
\itemize{
  \item Joshua Kunst \email{jbkunst@gmail.com}
}

Other contributors:
\itemize{
  \item Reto Wyss \email{reto.wyss.ch@gmail.com} [contributor]
  \item Ogan Mancarci \email{ogan.mancarci@gmail.com} [contributor]
}

}
\keyword{internal}
