% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchess.R
\docType{package}
\name{rchess}
\alias{rchess}
\alias{rchess-package}
\title{Package por chess move, generation/validation, piece placement/movement, and
check/checkmate/stalemate detection}
\description{
Package por chess move, generation/validation, piece placement/movement, and
check/checkmate/stalemate detection
}

