% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggchessboard.R, R/theme_null.R
\name{theme_null}
\alias{theme_null}
\title{Theme null}
\usage{
theme_null()

theme_null()
}
\description{
A ggplot2 theme null for ggchessboard

A ggplot2 theme null for ggchessboard
}

