% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chess.R
\docType{data}
\name{Chess}
\alias{Chess}
\title{Chess Class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000ae00c28> 
 - attr(*, "name")= chr "Chess_generator"
}}
\usage{
Chess
}
\description{
Chees class.
}
\keyword{datasets}

