% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chessboardjs.R
\name{chessboardjs}
\alias{chessboardjs}
\title{#' Plot a chessboard via chessboardjs}
\usage{

  chessboardjs(fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1",
  width = 300, height = 300)
}
\description{
#' Plot a chessboard via chessboardjs
}

