% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chessboardjs.R
\name{chessboardjsOutput}
\alias{chessboardjsOutput}
\title{Widget output function for use in Shiny}
\usage{
chessboardjsOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

