% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessboardjs.R
\name{chessboardjsOutput}
\alias{chessboardjsOutput}
\title{Widget output function for use in Shiny}
\usage{
chessboardjsOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Id of the div tag}

\item{width}{Width in pixels}

\item{height}{Height in pixels}
}
\description{
Widget output function for use in Shiny
}

