% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-chessopenings.R
\docType{data}
\name{chessopenings}
\alias{chessopenings}
\title{Chess openings data}
\format{A \code{data frame} with 544 observations and 3 variables.}
\usage{
chessopenings
}
\description{
This data comes from the chess Chess Opening Theory wikibook \url{https://en.wikibooks.org/wiki/Chess_Opening_Theory}.
The oppening were parse using \code{rvest} package.
}
\section{Links}{


\itemize{

 \item https://en.wikibooks.org/wiki/Chess_Opening_Theory/ECO_volume_A
 \item https://en.wikibooks.org/wiki/Chess_Opening_Theory/ECO_volume_B
 \item https://en.wikibooks.org/wiki/Chess_Opening_Theory/ECO_volume_C
 \item https://en.wikibooks.org/wiki/Chess_Opening_Theory/ECO_volume_D
 \item https://en.wikibooks.org/wiki/Chess_Opening_Theory/ECO_volume_E

}
}

\section{Variables}{


\itemize{

 \item \code{eco}: The game's opening classification
 \item \code{name}: Games's location
 \item \code{pgn}: The game's pgn

}
}
\examples{
data(chessopenings)
library("dplyr")
head(chessopenings)
}

