% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggchessboard.R
\name{ggchessboard}
\alias{ggchessboard}
\title{Plot a chessboard via ggplot2}
\usage{

  ggchessboard(fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1",
  cellcols = c("#D2B48C", "#F5F5DC"), perspective = "white",
  piecesize = 15)
}
\description{
Plot a chessboard via ggplot2
}
\examples{
\dontrun{
ggchessboard()

ggchessboard(fen = "rnbqkbnr/pp1ppppp/8/2p5/4P3/8/PPPP1PPP/RNBQKBNR w KQkq c6 0 2")

ggchessboard(fen = "rnbqkbnr/pp1ppppp/8/2p5/4P3/8/PPPP1PPP/RNBQKBNR w KQkq c6 0 2",
             cellcols = c("#CCCCCC", "#FAFAFA"),
             piecesize = 17,
             perspective = "black")
}
}

