% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rchess.R
\docType{package}
\name{rchess}
\alias{rchess}
\alias{rchess-package}
\title{rchess: A package por chess move, generation/validation, piece placement/movement, and
check/checkmate/stalemate detection}
\description{
a simpler code
}

