% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chessboardjs.R
\name{renderChessboardjs}
\alias{renderChessboardjs}
\title{Widget render function for use in Shiny}
\usage{
renderChessboardjs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{width}{Width in pixels}

\item{height}{Height in pixels}
}
\description{
Widget render function for use in Shiny
}

