% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessboardjs.R
\name{renderChessboardjs}
\alias{renderChessboardjs}
\title{Widget render function for use in Shiny}
\usage{
renderChessboardjs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression}

\item{env}{Environment}

\item{quoted}{Quoted}
}
\description{
Widget render function for use in Shiny
}

